Faucet generation + faucet website hosting
This adds a mechanism to generate different faucet accounts for
every teztnet, so we can stop relying on this large json file with 30000
faucet accounts that has been shared between all tezos testnets since
the beginning (and which recently got exhausted, forcing Nomadic to
reset the counter to zero)

Twin tezos-k8s PR: https://github.com/tqtezos/tezos-k8s/pull/224

The generation of faucet is deterministic and seeded by a github secret
concatenated with the full teztnet name. Accounts are thus unique per
teztnet.

We add a submodule to a fork of tezos-faucet repo on gitlab. It has been
modified to contain a chart and a deterministic account generation
script.

The tezos-k8s PR contains the same script, run on chain activation, so
the same accounts are funded.

This way, we pass the secret to both faucet website chart, and tezos
private chain chart, and they both re-generate the same list of accounts
independently. There is no need for a mechanism to pass large json files
between the charts.

We also import logic from tqinfra related to aws load balancer
controller. The faucet is an https ingress, and pulumi will error out if
I don't import this logic of selectively ignoring validation (not
totally clear on it, Aryeh wrote the code).

What is not done:

* handling the route53 alias. This likely requires k8s' route53
  controller
* handling https for faucet
* add a mechanism to start the faucet website with an account index
  larger than 0. If the infra is destroyed and recreated, you don't want
  to give again a faucet address that's already been used. You should
  optionally be able to pass a faucet_start_index variable
* modify the faucet frontend webpage to display which teztnet this
  faucet is for

But since I already deployed this PR from laptop, I recommend it to be
merged.