Step 2: create new network, but don't activate it yet
This should be done 12-24 hours before actication.

This creates a brand new network, with a new network name, but the
network will not activate. It will be stuck at block zero, but ingresses
will already be up and it will be already visible from the webpage.
Critically, the `--network https://teztnets.xyz/idiazabalnet` will have
the new network parameters.

Note that the faucet will not come up when the chain is not activating,
so to correct that, I set number of faucet accounts to zero. That must
be changed back to 10k when we actually start the network. Activating
the network with the wrong number of faucets is difficult to revert and
requires assistance from nomadic.

After this PR is merged, get in touch with Pierre Boutillier (kaml.fr),
Rodrigo Quelhas (smartpy) and Gino Imbrailo (ECAD) to restart their
bootstrap endpoints so they help bootstrap the network.

At this point, you should have a network with several peers, but no
blocks.

I recommend to do a "up" locally and the push to main and do a release.
Release is needed to update the `--network` endpoint and the entire
jekyll teztnets.xyz page with the new info.